astscript-radial-profile: Removed special sn option
Until now the radial profile script had a special option for the
signal-to-noise ratio (S/N)! However the S/N is already a "--measure" that
is provided by MakeCatalog. Having a separate option only for this column
made the script hard for the users to use (an extra option), and for the
maintainers to maintain (complexities in variable checking).

With this commit, this option has been removed, and within the manual the
user is instructed to use '--measure=sn' if they want the S/N. During the
process and while inspecting the code, I also made these corrections:

 - A new '--zeropoint' option has been added to the radial profile script
   so users can ask for things like the magnitude or surface brightness.

 - In the part where we converted WCS coordinates into pixel coordinates,
   the "sed 's/,/ /'" was extra (because Gnuastro programs, including
   'asttable' in this case, automatically separate columns by a comma).

 - With this commit, the '--stdhdu' option is optional (it is no longer
   used to distinguish a file from a number). To see if the value of
   '--stdhdu' is a file, we simply use 'if [ -f $stdhdu ]' (since it should
   exist anyway).

 - The 'tests/during-dev.sh' script (that is used for testing Gnuastro's
   programs and scripts) has been modified by including all the programs'
   configuration files during the build, and also including the place of
   all the built programs into the PATH. In this way, when testing an
   'astscript' with 'tests/during-dev.sh', we can be sure that the
   currently under-development programs will be used by the 'astscript-*'.