TAB completion: disabled directory removal due to bug
Until now, when a different directory has multiple files with the same
base-name, the TAB completion will delete the already-typed directory name
and would only print the matching filename (that may not exist in the
running directory). An example to reproduce the problem is given in the
comments above the code.

To let the TAB completion work for now, with this commit, I have disabled
the part that removes the directory name and we now just simply print the
full file name of matching files. This temporary fix is very ugly (because
long directories get repeated and is not the standard behavior of TAB
completion), but atleast it doesn't delete anything!