Crop: print message when image and WCS have differing dimensions
Until now, when the number of dimensions in the input image differed from
the number of dimensions of the WCS (with the WCS having a larger number of
dimensions) and WCS-based '--center' was given, Crop would crash with a
"segmentation fault"! This happens in some Radio data pipelines where a 2D
image has 4 WCS dimensions (the last two having zero length!).

This was happening because we were reading the values to '--center' based
on the number of dimensions in the WCS (while it was originally checked
with the number of dimensions in the image). Therefore when the number of
dimensions in the WCS was larger than the image, the program would attempt
to read from space that was not allocated, causing a segmentation fault.

With this commit, we now prepare the center positions based on the image
dimensions (the same dimensions that we check with). Therfore when the
dimensions differ, instead of an enigmatic "segmentation fault", a clear
warning is printed by Gnuastro's WCS library that the number of dimensions
differ. Its up to the user to fix this.

This fixes bug #60619.