Library (wcs.h): limiting header keys for wcs works with comment keys
Until now, when the input HDU to the 'gal_wcs_read_fitsptr' function had
empty keyword lines, or COMMENT or HISTORY keywords, the values given to
'hstartwcs' and 'hendwcs' would not be correct (the values to similarly
named options in the Crop and Warp programs! This was happening because we
were always setting CFITSIO's 'fits_hdr2str' function to ignore comments
when reading the keywords into a string!

With this commit, the second argument to CFITSIO's 'fits_hdr2str' function
is now optional: so when the user wants to actually give header keyword
numbers to ignore, the keyword numbers correspond to the same order they
see on the command-line with the 'astfits' program. When the user doesn't
want to limit the keywords, it will ignore comment and empty lines like
before (and not waste CPU power on extra lines).

This fixes bug #60634.