MakeCatalog: Brightness error correctly calculated when noise STD<1
Until now, the correction to fix error measurements in images with STD<1
was not being applied to the brightness error! It was applied when
calculating the signal-to-noise ratio (and thus the magnitude error), but
not the brightness error. As a result, when the data are in very small
units (like Janskies!) the brightness errors become larger than the
brightness itself, even for very high S/N objects!

This is explained more fully in Section 3.3 (S/N Equation Modiﬁcations) of
the NoiseChisel paper: when the image noise has a value that is less than
1, it means that the data are not in units of counts (they are often scaled
to other units, like electrons/sec). However, even though it is scaled
(multiplied by a factor) from the original counts, the original noise was
Poisson noise, it just needs to be scaled back.

With this commit, this multiplication factor is now applied to the
brightness error directly and thus propagating correctly to the S/N and the
magnitude error.

This fixes bug #60644.