NEWS: added names of people who contributed to features
Recently, Gnuastro has started the convention of mentioning the names of
people who contributed to a certain item in the NEWS file. But this was not
done systematically (I had just done this based on memory!).

With this commit, I went through all the commits over the NEWS since
Gnuastro 0.14 and added names mentioned in the commit message, when they
were missing in the NEWS file.