Library (arithmetic): randomly select from uniform distribution
Until now, the arithmetic library (and thus the arithmetic program, or
column arithmetic in Table) only had two ways to select a random value: a
Gaussian or Poisson distribution. But these distributions have a peak
value, so the selected values won't be uniformly distributed (which is
necessary in some scenarios).

Witht his commit, a new 'mknoise-uniform' operator has been added to allow
easy creation of random values that are uniformly distributed within the
requested range.