Library (arithmetic): differing envseed for each call to mknoise-*
Until now (since previous commit), if the 'mknoise-*' operators were needed
multiple times (in different columns for example), and a reproducible
result was required with '--envseed' all the invocations of 'mknoise-*'
operators would get the same random number generator seed!

With this commit, this problem has been fixed by incrementing the user's
random number generator seed for every call to the 'mknoise-*'
operators. As described in the manual, this isn't perfect, so Task 15971
has been defined as a possible solution (when there are multiple columns
and the order should not be important).

[1] https://savannah.gnu.org/task/?15971