Library (arithmetic.h): mag-to-counts is a new operator
Until now we had the 'counts-to-mag' operator in the arithmetic library
(and thus the Arithmetic program or column arithmetic in Table). But in
some scenarios (like adding Poisson noise to a mock image with a background
value in magnitudes), we need the inverse operator.

With this commit, the inverse operator has been added for this job and
besides its own documentation and example, an example of using it has been
added to the 'mknoise-poisson' operator.