NoiseChisel and Statistics: warning on bad --meanmedqdiff values
Until now, there was only a warning for a bad value to this option with the
Statistics program and it was missing in NoiseChisel. Furthermore, the
error message in Statistics didn't really describe the problem sufficiently
for a user who doesn't know that this option is.

With this commit, if the given value is larger than 0.5, NoiseChisel or
Statistics will crash with a very descriptive error message. Also, if the
value is larger than 0.3, a warning is printed, informing the user that
this is not a natural value and may be too much.