Book: corrected description of MakeCatalog's --halfsumradius
Until now, the description of '--halfsumradius' mentioned the old
'--halfradiusarea' option which is no longer available. Also it talked
about an 'obs' suffix (which its current name doesn't have anymore!).

With this commit, it now asks to look into the description of
'--halfsumarea' for more on teh potential problems and it doesn't talk
about an 'obs' suffix. The last paragraph is just a caution that this is
not equivalent to the effective radius.

This issue was reported by Samane Raji.