Book: notices in TAB completion section which is not yet complete
Currently TAB completion is not yet fully implemented in all the programs
and there are known issues in those that do.

With this commit, a notice has been added at the start of the respective
sections in the book to highlight this and let users know about potential
problems before they activate it.