Book: added description of gal_units_mag_to_counts
Until now, this function existed in the source of Gnuastro, but I had
forgot to include it in the book! After making the 0.15 release and on a
random check, I noticed that this function is missing!

With this commit, the description of this function has been added to the
book.