Author names: corrected multiple entries for Carlos
Carlos Morales-Socorro had made his commits with a slight difference
between the names: on one, his surname didn't have a '-'. As a result, in
the list of authors that is printed at the start of the PDF book (and in
the 'AUTHORS' file, his name was repeated two times.

With this commit, a Git mailmap entry has been added to correct this issue.

Also, I noticed a small typo in the "Why C?" section of the book (an 'of'
was mistakenly written for an 'or').