Book: corrected name of unit conversion library
By default, the node name of all the gnuastro libraries was just the simple
title (without the '*.h' header file name). The '*.h' header file name is
in the title of the page/section, but not in the node. However, by looking
at the new 0.15 release, I just noticed that this convention was not
followed for the units library (such that the node name had the parenthesis
and header name!).

With this commit, this has been corrected and the node name doesn't have
parenthesis or the header's name.