MakeCatalog: comments added for --halfsumsb and other columns
Until now, when the '--halfsumsb' column was requested from MakeCatalog
that also includes measuring clumps, the output column didn't have any
comments in the metadata. While looking at the code, I found that the
source of the problem is that the 'ccomment' (clump comment) variable was
not being set for this column!

With this commit 'ccomment' is given the same value as 'ocomment' (object
comments) like all the other columns that have the same description between
objects and clumps. After doing so, I inspected all the existing columns
and noticed that this was happening for a few more columns also: '--fwhm',
'--halfmaxradius', '--halfsumradius', '--fracmaxradius1',
'--fracmaxradius2', '--fracmaxsum1', '--fracmaxsum2'.

In the process, I also improved the description of '--halfsumsb' and
'--halfmaxsb' to be more clear (they didn't contain the term "Surface
brightness"!