Book: radial profile example using 0.2 as the tolerance for --sigmaclip
Until this commit, in an example describing how to use the radial profile
script, instead of using a tolerance it was a large number of iterations
(10).

Since it is better to have a tolerance value, with this commit the
termination criteria has been changed to 0.1 and the multiple of sigma has
been increased to 5 (to be sufficiently different from the default values
of 3,0.2).