Book: building now copies of the original .mailmap
Until now, for building the list of authors in the book we would make a
symbolic link to the main Gnuastro '.mailmap' file. However, apparently new
versions of Git don't support symbolic links, so it wasn't being taken into
affect!

With this commit, instead of the symbolic link, we are copying the file
from the top source directory temporarily (even the symbolic was being
deleted before this).