Radial profile: standard deviation input properly passed to MakeCatalog
Until now, the standard deviation input image was considered in the section
that inputs data sets cropped in the asked radius. However, the variable
that defined its usage for the MakeCatalog call was not given a value! As a
result, the standard deviation image would not be used in the catalog
production and all error-related measurements like the magnitude or
brightness errors or signal-to-noise ratio would be NaN.

With this commit, for all possibilities of having a standard deviation (as
a single number or file), a value has been defined for the standard
deviation cropped/warped file or the single digit.

This fixes Bug #60776.