MakeCatalog: brightnesserr is NaN when instd under label is blank
Until now, when the standard deviation image was NaN over the label's
pixels, but the values image is non-NaN, the brightness error column would
be zero. However, Zero is a meaningful number (for example in a mock image
when there is no error at all). So in the scenario above, the returned
value should be NaN, not zero.

With this commit, a new internal calculation column has been added for
counting how many standard deviation pixels were used. The returned
brightness error will be NaN when this number doesn't correspond to the
total number of values pixels pixels, or the values pixels are all
blank. In other words, the value will only be zero when the user explicitly
set those pixels in the standard deviation image to zero.

While doing this, I also noticed that the '--clumpsmagnitude' column was
asking for too many extra internal calculation columns that it didn't need!
So they have been removed.

This bug was found by Zahra Sharbaf.

This fixes bug #60778.