Fits: warning printed with --pixelscale and +4 dimensional data
Until now, when Fits was called with the '--pixelscale' option, the
area/volume of pixels in the input dataset was only printed when the
dataset had 2 or 3 dimensions, followed by a new-line. As a result, when
there were more than three dimensions, no new-line would be printed,
causing problems when it was called in a loop (outputs from different files
would be printed on the same line).

With this commit, when there are more than three dimensions and
'--pixelscale' is called with '--quiet', it will only print the pixel
scales and a new line (no volume or area information will be printed).