Book: fix to known issue in CFITSIO installation explained
On some operating systems, the 'testprog' program of CFITSIO will give a
linking error with the just-installed library, and crash. To fix it, it is
necessary to add the running directory to 'LD_LIBRARY_PATH'.

With this commit, an explanation has been added after the series of
commands to install CFITSIO, describing the problem and how to fix it.

This issue, and its fix, was reported by Aaron Watkins.