Arithmetic: new operator to find bounding box of an ellipse
Until now, Gnuastro only had a library function to find the width and
height of the bounding box of an ellipse (such that the box and ellipse
share the same center point). This function is used heavily in several
Gnuastro programs for low level operations, for example
MakeProfiles. However, there was no easy command-line interface to it.

With this commit, a wrapper for it has been written in the Arithmetic
library, allowing easy access to the command-line. It can work on both
table columns and images.

It was the first image-operator that returns two datasets, so I had to add
the new 'operands_add_list' function in Arithmetic's operands
functions.

During the testing, I also noticed a bug in the 'tofile-' and 'tofilefree-'
operators of Arithmetic (they weren't deleting an existing file, just
adding their dataset as new HDUs). So with this commit, that bug is also
fixed.

This fixes bug #60826.