Match: now possible to append non-matching rows of second table
Until now, when the input tables may have had repetative objects, the
process of merging two catalogs first required calling Match with
'--notmatched' to separate the objects of the second catalog that don't
match, then append the non-matching rows to the rows of the first. But this
extra was annoying/buggy! It would have been much more easier to do this
internally and in a single call to Match.

With this commit, we use the existing '--outcols' option and allow it to be
called with '--notmatched'. When they are called together like this, all
the first input's rows will be included in the output, but the non-matching
rows of the second input will also be added (for the desired columns).

To allow this, it is necessary that both inputs have the columns given to
'--outcols' and that each column has same numeric data type as the
corresponding column in the other catalog.