Book: added example to convert image pixels to surf.brightness
Until now, the only example to the 'counts-to-mag' operator of Arithmetic
was using a single value (not an image). This could be confusing.

With this commit, the discussion on how the per-pixel surface brightness
limit relates to the noise level (to show the example command) has been
moved into the footnotes. Also, just above the command, we say that you can
do this on an image by simply replacing the '0.1' by an image name.

Also, as a more concrete example of applying 'counts-to-mag' on an image,
in the description of surface brightness, I added an example of how to
create a surface brightness image.

This was implemented after comments from Zahra Sharbaf.