Match: further correction on acknowledgment of new feature
I had forgot to thank Leslie Hunt for her contribution to the motivation
behind the new feature in Match (to append non-matching rows of second
catalog to first). She is now acknowledged as a contributor to this task
(along with Juan as mentioned in Commit e62d723b4da8) in the NEWS file.