Match: correctly using string columns with --notmatched --outcols
Until now, the newly added feature of appending non-matched columns of
second input to first input was working fine with numbers, but when given
string columns, it caused a segmentation fault. This was happening because
string columns have a separately allocated space for each row and in the
first implementation, I was mistakenly double-freeing them!

With this commit, string columns have been given a special attention in the
processing and the segmentation fault has been fixed.