Book: added example usage of box-around-ellipse Arithmetic operator
The newly added 'box-around-ellipse' operator will return two operands on
the stack (unlike most other operators that only return a single
operand). In some situations, users may want to treat each separately, but
if they aren't experienced enough, it may be confusing/impossible to do so.

With this commit, an example has been added to show how it is possible to
treat each output separately with the 'set-' operator.