Query: avoid segmentation fault with NED and no dataset
Until now, when NED was queried without any dataset, Query would crash with
a segmentation fault! But it should have printed a warning about the
available datasets and how to access them (like any other database).

With this commit, the source of the problem was found and fixed (it was
introduced after we added the 'extinction' dataset to NED).