Library (txt.c): stdintimeout value passed as seconds+microseconds
Until now, the value given to 'stdintimeout' was directly passed onto the
'tv_usec' variable (that stores the delay time in micro-seconds) and the
value of 'tv_sec' (that stores the delay time in seconds) was set to
zero. This was based on a wrong understanding of these values! I mistakenly
thought that only one of the two should be used. However, they are
complementary: the number of seconds to delay should be given to 'tv_sec'
and the number of extra micro-seconds should be given to 'tv_usec'. In
other words, 'tv_usec' should not be larger than 1000000 (microseconds). As
a result, increasing the value of '--stdintimeout' beyond one million was
not effective.

On the other hand, the default waiting time for input from standard input
(a pipe) was 0.1 seconds (or 100000 micro-seconds). This could cause
confusing error messages for some users when the previous command takes
long. Thus forcing the user to increase the value with the '--stdintimeout'
option manually.

In line with the issue above, Zahra Sharbaf reported a bug where the test
for the sort-by-night script during 'make check' failed, because of this
error. Unfortunately a normal user had no way of fixing this without
manually editing the sort-by-night script (which is not good).

With this commit, the issues above have been addressed: 'tv_sec' and
'tv_usec' are now properly set, the default value of --stdintimeout has
been increased to 1.5 seconds (or 1500000 microseconds). This is still
enough time for a normal user to see an error if there is a problem in the
previous command (its not too long). But will greatly reduce the number of
un-expected errors. Also, generally, all programs are set to avoid checking
for standard input when a file has already been given (only Convolve was
not using the 'gal_options_check_stdin' function which does this
automatically, but it now uses it). For ConvertType, since the first
channel may be from the standard input, its '--stdintimeout' has been set
to the old value of 0.1sec.

Finally, since the sort-by-night script may be given thousands of images
that may take even longer than the default '--stdintimeout' value, this
option has been added to this script and passed to all the programs called
by the script that need it.

This commit fixes bug #60901 that was reported by Zahra Sharbaf.