Library (wcs.h): added step to help in debugging errors
Until now, when a WCSLIB error came up the only hint of the cause was the
finally printed WCSLIB error message.

With this commit, a commented line has been added to the WCS library so if
necessary, we can activate more detailed error messages (tracking the error
string to the first function that caused it). By default it is commented to
avoid slowing down the process, but when something is not clear, we can
just uncomment it.