Library (wcs.h): access to WCSLIB's wcshdo() only within library
Until now, the Crop and MakeProfiles programs directly called the 'wcshdo'
function of WCSLIB (that is in charge of writing the FITS
keywords). However, in time we have found some peculiarities with the WCS
keyword writing, for example if there is a TPV distortion, the CD matrix
should be used or some cleaning of the written parameters (sometimes
happening due to floating point errors).

With this commit a new function has been added to the 'wcs.h' library of
Gnuastro: 'gal_wcs_write_wcsstr' and the only place that we call WCSLIB's
'wcshdo' within the whole of Gnuastro is within this function. This greatly
helps in modularity and unifying the output behavior of all programs.

In the same spirit, the final CD/PC matrix correction (after the keywords
were written into the FITS file) have been moved to
'gal_fits_key_write_wcsstr' and the check for seeing if a CD matrix should
be used has been moved into a function and used in the
'gal_wcs_read_fitsptr' function (which is the only place we read WCS in
Gnuastro).

The necessity for all this was found after noting that the 'wcstoimg'
operator in column arithmetic is giving a WCSLIB error when the image has
TPV distortion (bug #60909). After contacting Mark Calabretta, he kindly
noted that the problem is in not using the CD matrix with TPV. I then
discovered that the tests we had added for this were not actually being
used by the Crop or MakeProfiles programs. So instead of repeating those
checks in these two programs (which would certainly cause many headaches in
the future), I made the generic changes above and now we have a single
place for reading/writing the WCS that will account for all existing/future
corrections.

This fixes bug #60909.