CosmicCalculator: warning printed for very low redshifts
Until now, the cosmology calculations would be simply printed on the output
for a certain redshift. However, when the distance is less than about 30
Mpc, the pecular velocity of the object may be significant compared to the
hubble flow, making the reported measurements invalid.

With this commit, in such cases a warning will be printed, letting the user
know about the problem and that they should be careful when interpretting
the results.

In parallel, I also noticed that we were giving old (arXiv-based) ADS URL
for the Plank 2018 Cosmology Parameters papers. That has also been
corrected with the ADS link that includes the finally published status of
the paper.

This was suggested by Ignacio Trujillo.