MakeCatalog: new column to measure the error in surface brightness
Until now there was no specific column to measure the error in surface
brightness. Although it is very similar to the error in measuring the
magnitude, it includes an area term that should also be taken into account.

With this commit, two new columns are added to MakeCatalog:
'--spatialresolution' and '--sberror'. The first is the error in measuring
area (or the FWHM of the PSF) which should be specified by the user. The
second is the measured error in measuring surface brightness that is a
column in the output catalog.

This task was suggested by Zahra Sharbaf and Juan Miro.