Radial profile: new an undersampling option
Until now, to create an aperture image and do the calculation, the actual
pixel size of the image was just considered.

With this commit, one block of code has been added, which allows the user
to define the bin size as the number of pixels and do the calculation
through the rebinned image.