MakeProfiles: new azimuthal angle profile
In order to make certain features (that only cover a certain angular
range), it was necessary to have every pixel's azimuthal angle (along the
ellipse of fixed radius that covers it). For example if it is necessary to
mask a certain arc over an image, or to build tidal tails or tidal shocks
over mock galaxies.

With this commit, a new function has been added to MakeProfiles that is
called 'azimuth' (or the code of '9'). It will return the azimuthal angle
for every pixel (in degrees).

While I was testing this new feature, I noticed that the function returning
the size of a box (in pixels) containing an ellipse doesn't can miss one
pixel along the length if its fractional value is more than 0.5 due to the
conversion of floating point to integer. This has also been corrected with
this commit, for both 2D ellipses and 3D ellipsoids.

This feature was added after some very productive discussions with Fernando
Buitrago and Matthias Kluge.