Radial profile: undersampling the apertures before filling the 0 pixel
Until this commit, the filling of the central pixel was done after the
undersampling of the aperture image. However, in order to make the script
more general and simpler, with this commit, this step has been changed. Now
the filling of the central pixel (0 value in the aperture image) is done
after the undersampling (re-binning) of the aperture image. If the
undersampling is not requested, then just a symbolic link is created.  In
addition to this, some extra information have been added to the script and
the Book in order to describe in a better way some steps.