Radial profile: subtracting 1 pixel to have the centered radius at 0
Until this commit, when undersampling or oversampling the radial profile,
the output radial profile started at Radius = 1. However, in order to have
the central value at Radius = 0, it is necessary to subtract to all radii
the value 1. This was properly done when no undersampling nor oversampling
was requested. With this commit, this bug has been corrected. Now, all the
output radial profiles start at Radius = 0 (central pixel value). In
addition to this, the option -z has been added to be able to use the short
version of --zeropoint.