Library (statistics.h): std function accounts for floating point errors
Until now, it could occur that the square of the average value was bigger
than the average of the squares of the values, due to all the elements in
the distribution having an almost identical value (all elements differing
by floating point errors). Hence, we would have a negative value under the
square root of the standard deviation calculation, returning a NaN.

With this commit, an extra check was added to 'gal_statistics_std' and
'gal_statistics_mean_std' to account for this: when the square of the
average value is bigger than the average of the square then the returned
value will be 0.0.

This fixes bug #60923.