Book: new section on precedence of Table operations
Until now there was no official place to see which table operation is done
before/after which operation. Describing this under each option was hard
and could be confusing.

With this commit, a new section has been added in the book to give a nice
general view on the various operations that can be done and also show which
one happens before/after which. The logic/reason behind the precedence is
also explained. The section finishes with an example command and a
description of how Table interprets it.

In the process (while checking the precedence on a test table), several
other issues were done:

 - I noticed that during column arithmetic we aren't reseting the internal
   flags of each column. As a result, the '--noblank' would now work in
   some scenarios. So with this commit, upon finishing column arithmetic
   the internal flags of the column are reset to zero.

 - Column concatenation has been moved before all row-selection or reording
   operations. This is much more logically and practically useful.

 - In the book, the description of reverse polish notation section has been
   greatly improved and edited to help new users.