PSF stamp: new script to generate star stamps and construct a PSF
This is the first commit of the new script: 'psf-stamp'. The main goal of
this script is to generate star stamps that can be stacked together in
order to construct a PSF. To do that, the script consider several
parameters. In short, it uses a center coordinates to crop the star with a
specified width. After that, if the user provides a segmentation map, it
will be also cropped in order to mask all the objects but the central one.
In this way, all contaminants will be masked according to that segmentation
image. Finally, a radial profile is computed to obtain the normalization
radius within a region specified by the user.  At the end, the output stamp
is masked (if the user has specified a segmentation image) and normalized.

It works on individual objects, not over a catalog of objects. Note that
the cropping of different stars from a catalog are independent tasks. So,
it is trivial to use this script with a for-loop to iterate over the entire
catalog, or alternatively, use Make to make the generation of the stamps in
paralell.