PSF select stars: new script for selecting stars
Until now, for building a non parametric psf, just the stamp of stras for
making different part of PSF was built, but at first for making different
part of psf we need to select stars in different range of magnitude and
then based on that make a stamp of stars in different range of magnitude
and stacked them together for making a PSF.

With this commit, for selecting stars in different part of magnitude, at
first with help of 'astquery' we download the stars from gaia
database. After downloading stars, the stars with good parallax have been
selected. In the next step for finding stars with axisratio above one, at
first by use of 'astmkcatalog' catalogue of clumps have been made and then
with use of 'astmatch' match this catalogue with catalogue of stars with
good parallax. Finally the stars with more than nine neighbors in special
distance have been deleted.

The changes in the book are due to 3251b5d2 commit in the branch of
master. It is in this commit due to the rebase. But this is not a problem
because those changes already exist in the master.

This script was originally written by Carlos Morales-Socorro.