PSF stamp: Renaming this script to 'psf-create-make-stamp'
Until this commit, the name of the script for creating the stamps was
'psf-stamp'. However, in order to have a good name convention for this and
future PSF scripts, I have renamed it. Now, all scripts related with the
creation of the PSF will be named as psf-create-* In the future, scripts
for making the subtraction of stars will be psf-subtract-*. This cames
after a discussion with Mohammad Akhlaghi and Sepideh Eskandarlou.