PSF stamp: fixing wrong directory of this script
In a previous commit in which I renamed this script, I mistakenly put it
into a wrong directory. This was just because I forgot to put the correct
directory path when moving with 'git mv ...'. With this commit, I am moving
the script to its proper location: bin/script/ sub-directory.