PSF model: new script for putting a PSF image into a sky position
With this commit, a new script has been added. It is desired to model the
scattered light field (from point-like sources) of an astronomical image.
To do that, it reads an input image, a PSF image, a center coordinates, and
a flux factor. With these inputs, the script will situate the PSF image
(scaled by the flux factor) into the center position provided.

Necessary offsets and cropping factors are computed internally. The final
output is the scaled PSF at the specified position with the same size than
the original input image. Note that this script works on invididual stars
and the flux factor has to be provided. As a consequence, to compute an
entire scattered light field model from several stars, it is necessary to
sum the different outputs of this script.