PSF model: new script for computing the PSF flux factor
With this commit, a new script has been added. It is desired to compute the
flux factor that is necessary to consider when modeling a particular star
using a PSF image.  Since a PSF image is arbitrary normalized, in order to
model a given star with the appropiate flux, it is necessary to scale
up/down the PSF image by a flux factor. This script is in charge of
computing that flux factor.

To do that, it reads an input image, the center of a star, a PSF image, and
a range of radii (and other parameters).  With these inputs, the script
will crop the star and compute its radial profile. In the same manner, the
radial profile of the PSF is also generated. After that, the ratio of the
two radial profiles is computed within the specified range of radius. The
flux factor is finally computed as an averaged value of these ratio
profile.