Library (fits.h): quiet option added to gal_fits_key_write_filename
Many of the programs write the filename of their inputs as keywords in the
output FITS files. When the filename is longer than 68 characters, they
would also print a warning that the full name is trunctated. However, this
warning did not account for the '--quiet' option! So even when the user
doesn't want any output printed on the command-line, the worning was
printed!

With this commit, the main function in charge of this operation
('gal_fits_key_write_filename') has been given a 'quiet' argument. When it
isn't zero, no warning will be printed. All the calls to this function
within all the programs have also been modified to take the value from the
user's input arguments.

This fixes bug #60958 that was reported by Raúl Infante-Sainz.