PSF model: checking that a PSF image is given for computing flux factor
Until this commit, the script that computes the flux factor was not
checking if a PSF image was given. This was causing the script to crash and
it took to me a while to notice that error.  With this commit, I have added
a check to ensure that a PSF image is provided with the option '--psf'. If
the user does not provide it, the script will crash and warn about this
problem with a good message.