PSF model: fixed bug in defining the radial profiles for flux factor
Until this commit, in the calculation of the flux factor there was a
problem in the definition of the radial profile names. The problem was that
the radial profile of the PSF and the object were not set properly. As a
consequence, the flux factor was not properly computed. With this commit,
this bug has been fixed. Now, both radial profiles are set properly and the
flux factor is the expected one.