PSF junction: new script for joining two PSF images
With this commit, a new script for joining two different PSF images has
been added. In general when making a very extended point spread function
(PSF), it is necessary to consider very bright stars. However, the problem
with that stars is that they are saturated in the core. Consequently, to
obtain the center without saturation, it is necessary to consider fainter
stars. So, at the end it is necessary to join these two different parts
into a one single image PSF (not saturated in the center, and high signal
to noise in the outer part). This script makes such joining.

To make the junction, the script consider the input image to be the outer
part of the PSF. The core part is provided with the option --core/-c. It
assumes that both PSF images are centered. In addition to that, the user
has to provide:
 --radius: the radius at which the joining is done.
 --fluxfactor: the multiplicative flux factor by which the core is scaled
   to have it at the same level than the outer part.

There are also other not mandatory options like the axis ratio and position
angle in case the two images are requested to be joined not in a circular
shape.