Statistics: new option to show quantile of the mean
The quantile of the mean is a very good measure of skewness in a
distribution. However, until now there was no simple option in the
Statistics program to measure it: you needed to first ask for the mean in
one call to statistics, then use the '--quantfunc' option in a second call.

With this function, statistics now has the '--quantofmean' option to easily
let the users easily find the quantile of the mean in the middle of other
measures they may need, all in one command.