Library (txt.h): content check when there is no metadata
Until now, when the user gave a plain-text file with no metadata, the
'txt.h' library would just counts the tokens in the top line (to find
number of columns in whole table) and assumed the contents are double
precision floating point.

It did ignore lines that are commented (starting with a '#'), however, when
the first non-commented line contains words (or generally anything that is
not a floating point number), this would cause very confusing error
messages later in the programs (since the number of columns was not counted
from the actual data, but the un-commented English text of the first line
for example).

With this commit, the plain-text table reading library checks if the tokens
in the first non-commented line that doesn't have any metadata, is actually
a number or not. It will abort with a warning if a non-number is found. In
the error, it will print the problematic token, so the users notices the
cause of the problem and corrects their input.

In the process, the example in the documentation of 'gal_type_to_string'
was corrected to be more clear on how to feed the pointer.

This fixes bug #60999.